Protein Protein Interaction with STRING and KEGG Pathways
Performing pathway enrichment analysis is a logical and important next step after using DAVID for functional annotation because it helps identify the broader biological pathways and networks involved in the dataset, providing a more integrated understanding of the gene functions.
The KEGG mapper tool is used to perform Pathways Enrichment Analysis for this project. It determines if a particular pathway is overrepresented in your gene list compared to a reference set, indicating a significant biological impact. The list of significant DEGs is uploaded to the KEGG search bar with “hsa” or “homo sapiens 9906” reference and the resulting set of matched pathways reveal pathways that are highly significant to biomarker identification and pathways that lessly significant to biomarker identification and those pathways which can be completed eliminated which allows us to further filter through the 23 gene list.